refactor: Add app interface
