+++ cmd/main.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/free5gc/nef/internal/logger"
-	nefapp "github.com/free5gc/nef/pkg/app"
 	"github.com/free5gc/nef/pkg/factory"
+	nefapp "github.com/free5gc/nef/pkg/service"
 	logger_util "github.com/free5gc/util/logger"
 	"github.com/free5gc/util/version"
 	"github.com/urfave/cli"
@@ -76,10 +76,10 @@ func action(cliCtx *cli.Context) error {
 
 	nef, err := nefapp.NewApp(ctx, cfg, tlsKeyLogPath)
 	if err != nil {
-		return fmt.Errorf("New NEF err: %+v", err)
+		return fmt.Errorf("new NEF err: %+v", err)
 	}
 
-	if err := nef.Run(); err != nil {
+	if err := nef.Start(); err != nil {
 		return nil
 	}
