+++ internal/sbi/processor/processor.go
@@ -4,10 +4,13 @@ import (
 	nef_context "github.com/free5gc/nef/internal/context"
 	"github.com/free5gc/nef/internal/sbi/consumer"
 	"github.com/free5gc/nef/internal/sbi/notifier"
+	"github.com/free5gc/nef/pkg/app"
 	"github.com/free5gc/nef/pkg/factory"
 )
 
 type nef interface {
+	app.App
+
 	Context() *nef_context.NefContext
 	Config() *factory.Config
 	Consumer() *consumer.Consumer
